chore: product 타입 파일 디렉토리 한 곳에 정리 및 컨벤션에 맞게 수정 (#86)
