+++ apps/web/services/products/createProduct.ts
@@ -32,7 +32,7 @@ const createProduct = async (
 /**
  * 이미지 storage에 업로드 후 DB에 데이터 추가 함수
  */
-export const registerProduct = async (
+const registerProduct = async (
   productData: ProductFormData,
   images: File[]
 ): Promise<ProductCreationResponse> => {
@@ -74,3 +74,5 @@ export const registerProduct = async (
     throw new Error('알 수 없는 오류가 발생했습니다');
   }
 };
+
+export { registerProduct };
